menjin/requirements.txt
alembic==0.9.4
certifi==2017.7.27.1
chardet==3.0.4
Flask-Migrate==2.1.0
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.7
MarkupSafe==1.0
olefile==0.44
Pillow==4.2.1
pymssql==2.1.3
python-dateutil==2.6.1
python-editor==1.0.3
qiniu==7.1.4
redis==2.10.5
requests==2.18.3
SQLAlchemy==1.1.12
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
